+++ controllers/news.js
@@ -30,19 +30,9 @@ const getById = async (req, res, next) => {
   }
 };
 
-const getCommentsByNews = async (req, res, next) => {
-  try{
-    const id = req.params.id
-    const response = await newsService.getCommentsByNews(id);
-    res.status(200).json(response)
-  } catch (error) {
-    next(error)
-  }
-}
 
 module.exports = {
   create,
   remove,
   getById,
-  getCommentsByNews,
 };
