+++ repositories/news.js
@@ -14,17 +14,10 @@ const getById = async (id) => {
   });
 };
 
-const getCommentsByNews = async (id) => {
-  return await db.Comments.findAll({
-    where:{
-      novelty_id: id
-    },
-  })
-}
+
 
 module.exports = {
   create,
   remove,
   getById,
-  getCommentsByNews,
 };
